% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntaov.R
\name{ntaov}
\alias{ntaov}
\title{Neutrosophic Two Way ANOVA}
\usage{
ntaov(dt)
}
\arguments{
\item{dt}{is a data frame}
}
\value{
Neutrosophic ANOVA Table
}
\description{
Neutrosophic Two Way ANOVA
}
\examples{
y=c(4,5,3,9,11,8,15,12,14)
y1=c(6,7,5,11,14,10,17,13,16)
tr=c(1,1,1,2,2,2,3,3,3)
cek=c(1,2,3,1,2,3,1,2,3)
dt=data.frame(y,y1,tr,cek)
ntaov(dt)
}
